[#369] Move "manage user roles" into approved stakeholders and entities
